Initial import, first version of server
First version of the whois server code. Right now data received from
client and whois server isn't validated so it is still quite unsafe to
use. However, you can connect to it from a whois client and it should
work just fine. Theoretically should be able to handle multiple
simultaneous connections but that is untested.